packages/legacy_integration_describer/pyproject.toml
[project]
name = "legacy-integration-describer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mp",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "typer>=0.24.1",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.4",
    "ty>=0.0.20",
]

[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true

[tool.uv.sources]
mp = { git = "https://github.com/chronicle/content-hub.git", subdirectory = "packages/mp" }

[tool.uv]
package = true

[tool.ruff]
exclude = ["src", "tests"]

[project.scripts]
legacy-integration-describer = "legacy_integration_describer.cli:app"

